main: Allow some vcpus to be real time
Add an option to set some vcpus as real time. With a cooperative guest,
this allows for getting real time behavior for a subset of guest
threads.  The guest will use cpusets to ensure that only real time tasks
are run on the real-time CPU, and that vcpu thread is real-time on the
host leading to behavior close to host real-time threads.

Change-Id: I87ca46437d8b00d7ecfe532dd520d18f6ebc193d
Signed-off-by: Dylan Reid <dgreid@chromium.org>