devices: block: convert to async for the poll loop.
Using async from block will allow for the use of asynchronous IO to
improve block speed.
Using an atomic instead of a mut ref for the enabled state and a non-mut
ref for the timer makes sharing the state between futures possible. Both
the queue handling and the timer firing will need access to the timer
and the enable boolean.

TEST=run fio-test from the guest and check that psees are comparable to
the previous commit.

Change-Id: I26856bbefc664efedd169df1e75bd4805b95bd42