devices: block: convert to async for the poll loop.
Using async from block will allow for the use of asynchronous IO to
improve block speed.
Using an atomic instead of a mut ref for the enabled state and a non-mut
ref for the timer makes sharing the state between futures possible. Both
the queue handling and the timer firing will need access to the timer
and the enable boolean.

Enable multi queue paths, but limit to one queue for now. Multi queue
will be used once async IO is available to enable more requests from the
guest to be used.

TEST=run fio-test from the guest and check that psees are comparable to
the previous commit.
TEST=boot and run fio test with virtio block.

Change-Id: I26856bbefc664efedd169df1e75bd4805b95bd42