wip: queue mutex
Allow the queue to be accessed from several async contexts with an async
mutex. The eventfd had to be wrapped as well to avoid cloning it and
having tons of FDs open. The queue and the event can't be in a shared
struct because that will require multip0le mut borrows of the
MutexGuard:(

Change-Id: I649c6fcc83cc265cdc89c6fdc81e6f02338d4918