devices: queue: Allow peek/pop of a descriptor index
Allow users of the queue to get the index of the next descriptor without
creating a descriptor. This will allow for multiple descriptors to be
processed in parallel in different contexts. As the descriptor borrows
GuestMemory, they can't be sent to other async contexts. The index can
be sent to another context that has its own copy of GuestMemory.

This will allow a main context to handle getting descriptor indices and
distribute them to async contexts that handle them in parallel.  The
descriptor won't be re-used by the guest until it is returned to the
used queue so host/guest synchronization isn't affected
by this change.

TEST=virtio block tests pass after converting block to async and using
the descriptor index values.
unittests pass and now peek/pop use this path.

Change-Id: Ia7fd60486159f8b2e31ab1dce852b55937a6a21b