devices: Add ability to consume descriptor offsets
When switching to async usage of descriptor queues, it is more useful to
consume the offsets and lengths of the descriptor buffers instead of the
buffers mapped as virtio slices. Having the offset,length pair and the
guest memory it refers to is enough to initiate async ops.

The new version skips the non-offset read/write case as using the
existing "seek" offset of the file doesn't make sense with async IO.

Adding the new version duplicates code from the existing volatile
slice version, that can be removed when all users convert to the new
version.

TEST=duplicate all tests that run on the volatile slice version.

Change-Id: I0a5e50694f40217e5d6ac40f5fcb82944a4c117e