tempfile: Expose a `path` member
Allow the `Builder` to return the path being built.  There are cases
where a user only wants a temporary path without a directory of file
created at the path. One example is socket tests in vhost_user, they use
the returned path to create a unix domain socket.
This is functionality that will avoid users having to use `format!` and
the `rand` crate's alphanumerics to generate paths.

Change-Id: Ia5f2a1479a4a06313ad3e80b35f1f1ea93f3e0e9